create new mapchete.path module containing MPath etc
